# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeparr API",
    "version": "0.5.0",
    "description": "Keeparr tags what to keep and reports what can be deleted — it never deletes anything itself.\n\n**Authentication**: most endpoints use the `keeparr_session` cookie (set by logging in through the web UI). Endpoints marked with `X-Api-Key` security also accept the admin API key (Settings → General) in the `X-Api-Key` header — intended for automation (`/api/stats`, `/api/admin/jobs`).\n\nAll ids (`ratingKey`) are the media server's item ids (Plex ratingKey or Jellyfin/Emby item id)."
  },
  "servers": [{ "url": "/", "description": "This Keeparr instance" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "keeparr_session",
        "description": "Signed session cookie set by the web login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Admin API key from Settings → General."
      }
    },
    "schemas": {
      "MediaCard": {
        "type": "object",
        "description": "One movie/series card. Arr fields are null/absent when the title isn't matched in Sonarr/Radarr.",
        "properties": {
          "ratingKey": { "type": "string" },
          "title": { "type": "string" },
          "year": { "type": "integer", "nullable": true },
          "sizeBytes": { "type": "integer" },
          "sizeLabel": { "type": "string" },
          "thumbUrl": { "type": "string", "nullable": true },
          "libraryKind": { "type": "string", "enum": ["movie", "show"] },
          "kept": { "type": "boolean", "description": "Protected — kept by anyone" },
          "keptByMe": { "type": "boolean" },
          "skipped": { "type": "boolean", "description": "This user's \"don't care\"" },
          "watched": { "type": "boolean" },
          "requestedByMe": { "type": "boolean" },
          "markedForDeleteByMe": { "type": "boolean" },
          "markedForDeleteAny": { "type": "boolean" },
          "source": { "type": "string", "enum": ["sonarr", "radarr"] },
          "quality": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "CleanupCampaignSummary": {
        "type": "object",
        "required": ["id", "name", "targetBytes", "deadlineAt", "gracePeriodDays", "graceEndsAt", "minScore", "status", "plannedItems", "plannedBytes", "reviewedItems", "reviewedBytes", "releasedItems", "releasedBytes", "protectedItems", "protectedBytes", "targetReached"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "targetBytes": { "type": "integer" },
          "deadlineAt": { "type": "integer", "description": "Review deadline (Unix seconds)" },
          "gracePeriodDays": { "type": "integer", "minimum": 0, "maximum": 90 },
          "graceEndsAt": { "type": "integer", "description": "Earliest close time (Unix seconds)" },
          "minScore": { "type": "integer", "minimum": 0, "maximum": 100 },
          "status": { "type": "string", "enum": ["active", "closed"] },
          "createdBy": { "type": "string" },
          "createdAt": { "type": "integer" },
          "closedAt": { "type": "integer", "nullable": true },
          "plannedItems": { "type": "integer" },
          "plannedBytes": { "type": "integer" },
          "reviewedItems": { "type": "integer" },
          "reviewedBytes": { "type": "integer" },
          "releasedItems": { "type": "integer", "description": "Reviewed and not protected by any live keep" },
          "releasedBytes": { "type": "integer" },
          "protectedItems": { "type": "integer", "description": "Snapshot titles currently kept by anyone" },
          "protectedBytes": { "type": "integer" },
          "targetReached": { "type": "boolean" }
        }
      },
      "CleanupCampaignItem": {
        "allOf": [
          { "$ref": "#/components/schemas/MediaCard" },
          {
            "type": "object",
            "required": ["rank", "score", "reasons", "reviewedByMe", "reviewCount", "protectedByAnyone", "outcome"],
            "properties": {
              "rank": { "type": "integer" },
              "score": { "type": "integer", "minimum": 0, "maximum": 100 },
              "reasons": { "type": "array", "items": { "type": "object", "properties": { "code": { "type": "string" }, "label": { "type": "string" }, "points": { "type": "integer" } } } },
              "reviewedByMe": { "type": "boolean" },
              "reviewCount": { "type": "integer" },
              "protectedByAnyone": { "type": "boolean" },
              "outcome": { "type": "string", "enum": ["pending", "released", "protected"] }
            }
          }
        ]
      },
      "CleanupCampaignDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/CleanupCampaignSummary" },
          { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/CleanupCampaignItem" } } } }
        ]
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/campaigns/{id}": {
      "get": {
        "tags": ["insights"],
        "summary": "Get a cleanup campaign snapshot and live outcomes",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Campaign detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "campaign": { "$ref": "#/components/schemas/CleanupCampaignDetail" }, "isAdmin": { "type": "boolean" } } } } } },
          "404": { "description": "Campaign not found" }
        }
      }
    },
    "/api/campaigns/{id}/review": {
      "post": {
        "tags": ["insights"],
        "summary": "Add the current user's release review",
        "description": "Allowed only while the campaign is active and through its review deadline. Any live household keep still overrides release.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["ratingKey"], "properties": { "ratingKey": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated campaign detail" }, "409": { "description": "Campaign/item is not reviewable" } }
      },
      "delete": {
        "tags": ["insights"],
        "summary": "Undo the current user's release review",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["ratingKey"], "properties": { "ratingKey": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated campaign detail" }, "409": { "description": "Campaign/item is not reviewable" } }
      }
    },
    "/api/admin/campaigns": {
      "post": {
        "tags": ["admin"],
        "summary": "Create an immutable Smart Reclaim campaign snapshot",
        "description": "Admin web session required. Snapshots every current unprotected candidate matching the optional sections and minimum score; the target is a progress goal, not a row limit.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "targetBytes", "deadlineAt"], "properties": { "name": { "type": "string", "maxLength": 100 }, "targetBytes": { "type": "integer", "minimum": 1 }, "deadlineAt": { "type": "integer", "description": "Future Unix timestamp" }, "gracePeriodDays": { "type": "integer", "minimum": 0, "maximum": 90, "default": 7 }, "minScore": { "type": "integer", "minimum": 0, "maximum": 100, "default": 45 }, "sectionIds": { "type": "array", "maxItems": 100, "items": { "type": "string" } } } } } } },
        "responses": { "201": { "description": "Campaign created" }, "403": { "description": "Admin required" }, "409": { "description": "No candidates match" } }
      }
    },
    "/api/admin/campaigns/{id}/close": {
      "post": {
        "tags": ["admin"],
        "summary": "Close a campaign after its grace period",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Closed campaign summary" }, "401": { "description": "Valid web session required" }, "403": { "description": "Admin required" }, "404": { "description": "Campaign not found" }, "409": { "description": "Grace period is still active or campaign is already closed" } }
      }
    },
...
  }
}